+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Use plone.volto uid_to_url method to convert resolveuid links in summary.
   [cekk]
-
+- Patch plone.restapi RESOLVEUID_RE regexp to catch more urls.
+  [cekk]
 
 5.2.0 (2023-08-21)
 ------------------