+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
   [cekk]
 - Patch plone.restapi RESOLVEUID_RE regexp to catch more urls.
   [cekk]
+- Ignore non-existing indexes in custom ranking.
+  [cekk]
+  
 
 5.2.0 (2023-08-21)
 ------------------